refactor: moving each storage target into their own module
