+++ dev/benchmarks/storage/native/benches/native_storage_engines.rs
@@ -1,6 +1,5 @@
-use std::collections::HashMap;
 use std::future::Future;
-use std::path::{Path, PathBuf};
+use std::path::Path;
 use std::pin::pin;
 use std::task::{Context, Poll, RawWaker, RawWakerVTable, Waker};
 use std::time::{Duration, Instant};
@@ -13,388 +12,24 @@ use criterion::{
     BenchmarkGroup, Criterion, Throughput, black_box, criterion_group, criterion_main,
     measurement::WallTime,
 };
-use redb::{
-    Database, Durability, ReadOnlyTable, ReadableDatabase, ReadableTable, TableDefinition,
-    WriteTransaction,
-};
-use rocksdb::{
-    BlockBasedOptions, Cache, DB, DBCompressionType, Direction, IteratorMode, Options, ReadOptions,
-    WriteBatch,
-};
-use tempfile::TempDir;
+#[path = "native_storage_engines/redb.rs"]
+mod redb;
+#[path = "native_storage_engines/rocksdb.rs"]
+mod rocksdb;
+#[path = "native_storage_engines/sqlite.rs"]
+mod sqlite;
+
+use self::redb::RedbEngine;
+use self::rocksdb::RocksDbEngine;
+use self::sqlite::SqliteEngine;
 
 const CACHE_BYTES: usize = 32 * 1024 * 1024;
-const SQLITE_CACHE_KIB: i64 = -(CACHE_BYTES as i64 / 1024);
 const DATA_DIR: &str = concat!(env!("CARGO_MANIFEST_DIR"), "/../data");
-const REDB_TABLE: TableDefinition<&[u8], &[u8]> = TableDefinition::new("kv");
 
 trait NativeEngine: BenchEngine + Sized {
     const NAME: &'static str;
 
     fn open_fresh() -> Result<Self, EngineError>;
-}
-
-struct SqliteEngine {
-    path: PathBuf,
-    conn: Option<rusqlite::Connection>,
-    _dir: TempDir,
-}
-
-impl SqliteEngine {
-    fn conn(&self) -> Result<&rusqlite::Connection, EngineError> {
-        self.conn
-            .as_ref()
-            .ok_or_else(|| EngineError::new("sqlite connection is closed"))
-    }
-}
-
-impl NativeEngine for SqliteEngine {
-    const NAME: &'static str = "rusqlite";
-
-    fn open_fresh() -> Result<Self, EngineError> {
-        let dir = tempfile::tempdir().map_err(eng)?;
-        let path = dir.path().join("bench.sqlite");
-        let conn = open_sqlite_conn(&path)?;
-        Ok(Self {
-            path,
-            conn: Some(conn),
-            _dir: dir,
-        })
-    }
-}
-
-impl BenchEngine for SqliteEngine {
-    fn put(&mut self, key: &[u8], value: &[u8]) -> Result<(), EngineError> {
-        self.conn()?
-            .prepare_cached("INSERT OR REPLACE INTO kv(k, v) VALUES(?1, ?2)")
-            .map_err(eng)?
-            .execute(rusqlite::params![key, value])
-            .map_err(eng)?;
-        Ok(())
-    }
-
-    fn get(&mut self, key: &[u8]) -> Result<Option<u8>, EngineError> {
-        let mut stmt = self
-            .conn()?
-            .prepare_cached("SELECT v FROM kv WHERE k = ?1")
-            .map_err(eng)?;
-        match stmt.query_row(rusqlite::params![key], |row| row.get::<_, Vec<u8>>(0)) {
-            Ok(value) => Ok(value.first().copied()),
-            Err(rusqlite::Error::QueryReturnedNoRows) => Ok(None),
-            Err(error) => Err(eng(error)),
-        }
-    }
-
-    fn delete(&mut self, key: &[u8]) -> Result<(), EngineError> {
-        self.conn()?
-            .prepare_cached("DELETE FROM kv WHERE k = ?1")
-            .map_err(eng)?
-            .execute(rusqlite::params![key])
-            .map_err(eng)?;
-        Ok(())
-    }
-
-    fn range(&mut self, lo: &[u8], hi: &[u8], limit: usize) -> Result<usize, EngineError> {
-        let mut stmt = self
-            .conn()?
-            .prepare_cached("SELECT 1 FROM kv WHERE k >= ?1 AND k < ?2 ORDER BY k LIMIT ?3")
-            .map_err(eng)?;
-        let rows = stmt
-            .query_map(rusqlite::params![lo, hi, limit as i64], |_| Ok(()))
-            .map_err(eng)?
-            .count();
-        Ok(rows)
-    }
-
-    fn begin_phase(&mut self, _kind: PhaseKind) -> Result<(), EngineError> {
-        self.conn()?.execute_batch("BEGIN").map_err(eng)
-    }
-
-    fn end_phase(&mut self, _kind: PhaseKind) -> Result<(), EngineError> {
-        self.conn()?.execute_batch("COMMIT").map_err(eng)
-    }
-
-    async fn reopen(&mut self) -> Result<(), EngineError> {
-        self.conn = None;
-        self.conn = Some(open_sqlite_conn(&self.path)?);
-        Ok(())
-    }
-}
-
-struct RocksDbEngine {
-    path: PathBuf,
-    db: Option<DB>,
-    pending: Option<PendingBatch>,
-    _dir: TempDir,
-}
-
-struct PendingBatch {
-    batch: WriteBatch,
-    overlay: HashMap<Vec<u8>, Option<Vec<u8>>>,
-}
-
-impl RocksDbEngine {
-    fn db(&self) -> Result<&DB, EngineError> {
-        self.db
-            .as_ref()
-            .ok_or_else(|| EngineError::new("rocksdb is closed"))
-    }
-}
-
-impl NativeEngine for RocksDbEngine {
-    const NAME: &'static str = "rocksdb";
-
-    fn open_fresh() -> Result<Self, EngineError> {
-        let dir = tempfile::tempdir().map_err(eng)?;
-        let path = dir.path().join("bench.rocksdb");
-        let db = open_rocksdb(&path)?;
-        Ok(Self {
-            path,
-            db: Some(db),
-            pending: None,
-            _dir: dir,
-        })
-    }
-}
-
-impl BenchEngine for RocksDbEngine {
-    fn put(&mut self, key: &[u8], value: &[u8]) -> Result<(), EngineError> {
-        if let Some(pending) = self.pending.as_mut() {
-            pending.batch.put(key, value);
-            pending.overlay.insert(key.to_vec(), Some(value.to_vec()));
-        } else {
-            self.db()?.put(key, value).map_err(eng)?;
-        }
-        Ok(())
-    }
-
-    fn get(&mut self, key: &[u8]) -> Result<Option<u8>, EngineError> {
-        if let Some(pending) = &self.pending {
-            if let Some(value) = pending.overlay.get(key) {
-                return Ok(value.as_ref().and_then(|bytes| bytes.first().copied()));
-            }
-        }
-        Ok(self
-            .db()?
-            .get(key)
-            .map_err(eng)?
-            .and_then(|value| value.first().copied()))
-    }
-
-    fn delete(&mut self, key: &[u8]) -> Result<(), EngineError> {
-        if let Some(pending) = self.pending.as_mut() {
-            pending.batch.delete(key);
-            pending.overlay.insert(key.to_vec(), None);
-        } else {
-            self.db()?.delete(key).map_err(eng)?;
-        }
-        Ok(())
-    }
-
-    fn range(&mut self, lo: &[u8], hi: &[u8], limit: usize) -> Result<usize, EngineError> {
-        if limit == 0 {
-            return Ok(0);
-        }
-
-        let mut read_options = ReadOptions::default();
-        read_options.set_iterate_upper_bound(hi.to_vec());
-        let iter = self
-            .db()?
-            .iterator_opt(IteratorMode::From(lo, Direction::Forward), read_options);
-
-        let mut rows = 0usize;
-        for item in iter {
-            item.map_err(eng)?;
-            rows += 1;
-            if rows == limit {
-                break;
-            }
-        }
-        Ok(rows)
-    }
-
-    fn begin_phase(&mut self, kind: PhaseKind) -> Result<(), EngineError> {
-        if kind.is_write() {
-            if self.pending.is_some() {
-                return Err(EngineError::new("rocksdb write batch is already open"));
-            }
-            self.pending = Some(PendingBatch {
-                batch: WriteBatch::default(),
-                overlay: HashMap::new(),
-            });
-        }
-        Ok(())
-    }
-
-    fn end_phase(&mut self, kind: PhaseKind) -> Result<(), EngineError> {
-        if kind.is_write() {
-            let pending = self
-                .pending
-                .take()
-                .ok_or_else(|| EngineError::new("rocksdb write batch is not open"))?;
-            self.db()?.write(&pending.batch).map_err(eng)?;
-        }
-        Ok(())
-    }
-
-    async fn reopen(&mut self) -> Result<(), EngineError> {
-        self.pending = None;
-        self.db = None;
-        self.db = Some(open_rocksdb(&self.path)?);
-        Ok(())
-    }
-}
-
-struct RedbEngine {
-    path: PathBuf,
-    db: Option<Database>,
-    write_txn: Option<WriteTransaction>,
-    read_table: Option<ReadOnlyTable<&'static [u8], &'static [u8]>>,
-    _dir: TempDir,
-}
-
-impl RedbEngine {
-    fn db(&self) -> Result<&Database, EngineError> {
-        self.db
-            .as_ref()
-            .ok_or_else(|| EngineError::new("redb is closed"))
-    }
-
-    fn active_write_txn(&self) -> Result<&WriteTransaction, EngineError> {
-        self.write_txn
-            .as_ref()
-            .ok_or_else(|| EngineError::new("redb write transaction is not open"))
-    }
-}
-
-impl NativeEngine for RedbEngine {
-    const NAME: &'static str = "redb";
-
-    fn open_fresh() -> Result<Self, EngineError> {
-        let dir = tempfile::tempdir().map_err(eng)?;
-        let path = dir.path().join("bench.redb");
-        let db = Database::create(&path).map_err(eng)?;
-        Ok(Self {
-            path,
-            db: Some(db),
-            write_txn: None,
-            read_table: None,
-            _dir: dir,
-        })
-    }
-}
-
-impl BenchEngine for RedbEngine {
-    fn put(&mut self, key: &[u8], value: &[u8]) -> Result<(), EngineError> {
-        let mut table = self
-            .active_write_txn()?
-            .open_table(REDB_TABLE)
-            .map_err(eng)?;
-        table.insert(key, value).map_err(eng)?;
-        Ok(())
-    }
-
-    fn get(&mut self, key: &[u8]) -> Result<Option<u8>, EngineError> {
-        if let Some(table) = &self.read_table {
-            return Ok(table
-                .get(key)
-                .map_err(eng)?
-                .and_then(|value| value.value().first().copied()));
-        }
-
-        let table = self
-            .active_write_txn()?
-            .open_table(REDB_TABLE)
-            .map_err(eng)?;
-        Ok(table
-            .get(key)
-            .map_err(eng)?
-            .and_then(|value| value.value().first().copied()))
-    }
-
-    fn delete(&mut self, key: &[u8]) -> Result<(), EngineError> {
-        let mut table = self
-            .active_write_txn()?
-            .open_table(REDB_TABLE)
-            .map_err(eng)?;
-        table.remove(key).map_err(eng)?;
-        Ok(())
-    }
-
-    fn range(&mut self, lo: &[u8], hi: &[u8], limit: usize) -> Result<usize, EngineError> {
-        if limit == 0 {
-            return Ok(0);
-        }
-
-        let count_rows = |table: &ReadOnlyTable<&'static [u8], &'static [u8]>| {
-            let mut rows = 0usize;
-            for item in table.range(lo..hi).map_err(eng)? {
-                item.map_err(eng)?;
-                rows += 1;
-                if rows == limit {
-                    break;
-                }
-            }
-            Ok(rows)
-        };
-
-        if let Some(table) = &self.read_table {
-            return count_rows(table);
-        }
-
-        let table = self
-            .active_write_txn()?
-            .open_table(REDB_TABLE)
-            .map_err(eng)?;
-        let mut rows = 0usize;
-        for item in table.range(lo..hi).map_err(eng)? {
-            item.map_err(eng)?;
-            rows += 1;
-            if rows == limit {
-                break;
-            }
-        }
-        Ok(rows)
-    }
-
-    fn begin_phase(&mut self, kind: PhaseKind) -> Result<(), EngineError> {
-        if kind.is_write() {
-            if self.write_txn.is_some() {
-                return Err(EngineError::new("redb write transaction is already open"));
-            }
-            let mut txn = self.db()?.begin_write().map_err(eng)?;
-            txn.set_durability(Durability::None).map_err(eng)?;
-            self.write_txn = Some(txn);
-        } else {
-            if self.read_table.is_some() {
-                return Err(EngineError::new("redb read table is already open"));
-            }
-            let txn = self.db()?.begin_read().map_err(eng)?;
-            self.read_table = Some(txn.open_table(REDB_TABLE).map_err(eng)?);
-        }
-        Ok(())
-    }
-
-    fn end_phase(&mut self, kind: PhaseKind) -> Result<(), EngineError> {
-        if kind.is_write() {
-            let txn = self
-                .write_txn
-                .take()
-                .ok_or_else(|| EngineError::new("redb write transaction is not open"))?;
-            txn.commit().map_err(eng)?;
-        } else {
-            self.read_table = None;
-        }
-        Ok(())
-    }
-
-    async fn reopen(&mut self) -> Result<(), EngineError> {
-        self.write_txn = None;
-        self.read_table = None;
-        self.db = None;
-        self.db = Some(Database::open(&self.path).map_err(eng)?);
-        Ok(())
-    }
 }
 
 struct DatasetView<'a> {
@@ -547,39 +182,6 @@ fn load_dataset(benchmark: &Benchmark) -> KvDataset {
         .unwrap_or_else(|error| panic!("decode fixture {}: {error}", benchmark.kv_fixture));
     assert_eq!(dataset.profile, benchmark.profile);
     dataset
-}
-
-fn open_sqlite_conn(path: &Path) -> Result<rusqlite::Connection, EngineError> {
-    let conn = rusqlite::Connection::open(path).map_err(eng)?;
-    conn.execute_batch(&format!(
-        "PRAGMA page_size = 16384;
-         PRAGMA journal_mode = WAL;
-         PRAGMA synchronous = NORMAL;
-         PRAGMA cache_size = {SQLITE_CACHE_KIB};
-         PRAGMA temp_store = MEMORY;
-         PRAGMA busy_timeout = 5000;
-         CREATE TABLE IF NOT EXISTS kv(
-             k BLOB PRIMARY KEY,
-             v BLOB NOT NULL
-         ) WITHOUT ROWID;",
-    ))
-    .map_err(eng)?;
-    Ok(conn)
-}
-
-fn open_rocksdb(path: &Path) -> Result<DB, EngineError> {
-    let mut block_options = BlockBasedOptions::default();
-    block_options.set_bloom_filter(10.0, false);
-    let cache = Cache::new_lru_cache(CACHE_BYTES);
-    block_options.set_block_cache(&cache);
-
-    let mut options = Options::default();
-    options.create_if_missing(true);
-    options.set_block_based_table_factory(&block_options);
-    options.set_compression_type(DBCompressionType::Lz4);
-    options.set_bottommost_compression_type(DBCompressionType::Zstd);
-
-    DB::open(&options, path).map_err(eng)
 }
 
 fn eng<E: ToString>(error: E) -> EngineError {
